+++ HierarchyDecorator/Scripts/Runtime/ExampleScript2.cs
@@ -7,7 +7,7 @@ public class ExampleScript2 : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        
+
     }
 
     // Update is called once per frame